loan and application finish

+++ app/Http/Controllers/Admin/ApplicationController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 
 class ApplicationController extends Controller
@@ -14,18 +15,34 @@ class ApplicationController extends Controller
     public function index()
     {
         // dd(Auth::user()->role->slug);
+        Gate::authorize('app.applications.index');
         if (Auth::user()->role->slug == 'super-admin') {
-            $applications = DB::table('loan_card_apply')->latest()->get();
+            $applications = DB::table('loan_card_apply')->where('type','loan')->latest()->get();
         }else {
-            $applications = DB::table('loan_card_apply')->where('send_to_vendor',Auth::user()->id)->latest()->get();
+            $applications = DB::table('loan_card_apply')->where('send_to_vendor',Auth::user()->id)->where('type','loan')->latest()->get();
         }
     //    dd($applications);
 
-        return view('admin.application.index', compact('applications'));
+        return view('admin.application.loan.index', compact('applications'));
+    }
+
+    public function cardIndex()
+    {
+        // dd(Auth::user()->role->slug);
+        Gate::authorize('app.applications.index');
+        if (Auth::user()->role->slug == 'super-admin') {
+            $applications = DB::table('loan_card_apply')->where('type','card')->latest()->get();
+        }else {
+            $applications = DB::table('loan_card_apply')->where('send_to_vendor',Auth::user()->id)->where('type','card')->latest()->get();
+        }
+    //    dd($applications);
+
+        return view('admin.application.card.index', compact('applications'));
     }
 
     public function edit($id)
     {
+        Gate::authorize('app.applications.edit');
         $application = DB::table('loan_card_apply')->where('id', $id)->first();
         $role = Role::where('slug','vendor')->first()->id;
         $vendors = Role::findOrFail($role)->users()->get();
@@ -35,36 +52,42 @@ class ApplicationController extends Controller
 
     public function update(Request $request, $id) {
         // return $request->all();
+        Gate::authorize('app.applications.edit');
+        $formData = DB::table('loan_card_apply')->where('id', $id)->first();
         $data = [
             'author_note' => $request->author_note,
             'vendor_note' => isset($request->vendor_note) ? $request->vendor_note : null,
-            'send_to_vendor' => $request->send_to_vendor,
+            'send_to_vendor' => isset($request->send_to_vendor) ? $request->send_to_vendor : $formData->send_to_vendor,
             'status' => $request->status
         ];
         // dd($data);
         $formUpdate = DB::table('loan_card_apply')->where('id', $id)->update($data);
-        return redirect()->route('admin.application.index');
+        return redirect()->back();
     }
 
     public function destroy($id) {
+        Gate::authorize('app.applications.destroy');
         DB::table('loan_card_apply')->where('id', '=', $id)->delete();
         return redirect()->back();
     }
 
     public function lifegeneralIndex()
     {
+        Gate::authorize('app.applications.index');
         $applications = DB::table('life_general_form')->get();
         return view('admin.application.life-general.index', compact('applications'));
     }
 
     public function lifegeneralEdit($id)
     {
+        Gate::authorize('app.applications.edit');
         $application = DB::table('life_general_form')->where('id', $id)->first();
         return view('admin.application.life-general.edit', compact('application'));
     }
 
     public function lifegeneralUpdate(Request $request,$id)
     {
+        Gate::authorize('app.applications.edit');
         // return $request->all();
         $data = [
             'admin_note' => $request->admin_note,
@@ -76,18 +99,21 @@ class ApplicationController extends Controller
     }
 
     public function lifegeneralDestroy($id) {
+        Gate::authorize('app.applications.destroy');
         DB::table('life_general_form')->where('id', '=', $id)->delete();
         return redirect()->back();
     }
 
     public function transportIndex()
     {
+        Gate::authorize('app.applications.index');
         $applications = DB::table('bike_car_form')->get();
         return view('admin.application.bike_car.index', compact('applications'));
     }
 
     public function transportEdit($id)
     {
+        Gate::authorize('app.applications.edit');
         $application = DB::table('bike_car_form')->where('id', $id)->first();
         return view('admin.application.bike_car.edit', compact('application'));
     }
@@ -95,6 +121,7 @@ class ApplicationController extends Controller
     public function transportUpdate(Request $request,$id)
     {
         // return $request->all();
+        Gate::authorize('app.applications.edit');
         $data = [
             'admin_note' => $request->admin_note,
             'status' => $request->status
@@ -105,6 +132,7 @@ class ApplicationController extends Controller
     }
 
     public function transportDestroy($id) {
+        Gate::authorize('app.applications.destroy');
         $old_file = DB::table('bike_car_form')->where('id', '=', $id)->select('file')->get();
         if (Storage::disk('public')->exists('frontend/assets/files/insurances/form/pdf/'.$old_file)) {
             Storage::disk('public')->delete('frontend/assets/files/insurances/form/pdf/'.$old_file);
